Move the SAML library into the Saml module, one type per file (#800)
Saml.cs was the last multi-type file left at the SSO-Auth project root — the
inherited AspNetSaml component holding two unrelated types (the inbound response
parser and the outbound AuthnRequest builder) in one 722-line file under the root
namespace. Fold it into the Saml module and give each type its own file, the
one-type-per-file rule the rest of the module already follows:

- SamlResponse.cs (git-renamed from Saml.cs — history preserved for the larger
  type): the inbound assertion parser/validator.
- SamlAuthnRequest.cs: the outbound redirect-binding request builder, with its
  using set trimmed to what it actually references.

Both move to namespace Jellyfin.Plugin.SSO_Auth.Api.Saml; consumers that reached
the types through the root namespace gain the module using. The upstream Apache
license header travels with both halves.

No behavior change. Verified by the full build (--warnaserror, both TFMs, 0
warnings), the complete suite (1588 both TFMs), and the ABI-floor build.

Refs #800
