+++ SSO-Auth.Tests/Saml/SamlCrlfInteropTests.cs
@@ -1,4 +1,5 @@
 using Jellyfin.Plugin.SSO_Auth;
+using Jellyfin.Plugin.SSO_Auth.Api.Saml;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
