+++ SSO-Auth.Tests/Saml/SamlSecondaryCertificateTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Xml;
 using Jellyfin.Plugin.SSO_Auth;
+using Jellyfin.Plugin.SSO_Auth.Api.Saml;
 using Xunit;
 
 namespace Jellyfin.Plugin.SSO_Auth.Tests;
